refactor : queryGenerator 구현체
