+++ src/main/java/core/jdbc/custom/query/QueryGenerator.java
@@ -1,4 +1,4 @@
-package core.jdbc.custom;
+package core.jdbc.custom.query;
 
 @FunctionalInterface
 public interface QueryGenerator {
